+++ proyectoLibreria/templates_c/85a40332c1ed75b13f5fd72740da410886a04353_0.file.brands.tpl.php
@@ -1,18 +1,18 @@
 <?php
-/* Smarty version 4.2.1, created on 2022-10-10 11:05:11
+/* Smarty version 4.2.1, created on 2022-10-11 10:58:35
   from 'C:\xampp\htdocs\tpe\proyectoLibreria\templates\brands.tpl' */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '4.2.1',
-  'unifunc' => 'content_6343e04764e591_19714775',
+  'unifunc' => 'content_6345303b7e15c0_49244351',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
     '85a40332c1ed75b13f5fd72740da410886a04353' => 
     array (
       0 => 'C:\\xampp\\htdocs\\tpe\\proyectoLibreria\\templates\\brands.tpl',
-      1 => 1665392710,
+      1 => 1665478709,
       2 => 'file',
     ),
   ),
@@ -22,7 +22,7 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
     'file:footer.tpl' => 1,
   ),
 ),false)) {
-function content_6343e04764e591_19714775 (Smarty_Internal_Template $_smarty_tpl) {
+function content_6345303b7e15c0_49244351 (Smarty_Internal_Template $_smarty_tpl) {
 $_smarty_tpl->_subTemplateRender("file:header.tpl", $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
 ?>
 
@@ -36,8 +36,10 @@ $_smarty_tpl->_subTemplateRender("file:header.tpl", $_smarty_tpl->cache_id, $_sm
         <tr>
             <th bgcolor='#BFBFBF' style="font-weight: bold;">Empresas</th>
             <th>Ver empresa</th>
-            <th>Eliminar empresa</th>
-            <th>Editar empresa</th>
+            <?php if ((isset($_SESSION['USER_ID']))) {?>
+                <th>Eliminar empresa</th>
+                <th>Editar empresa</th>
+            <?php }?>
         </tr>
     </thead>
     
@@ -61,16 +63,17 @@ $_smarty_tpl->tpl_vars['brand']->do_else = false;
 " type='button' class='btn btn-primary ml-auto'>VER</a>
                 </td>
 
-                <td>
-                    <a href='delete/<?php echo $_smarty_tpl->tpl_vars['brand']->value->id_brand;?>
+                <?php if ((isset($_SESSION['USER_ID']))) {?>
+                    <td>
+                        <a href='delete/<?php echo $_smarty_tpl->tpl_vars['brand']->value->id_brand;?>
 'type='button' class="btn btn-danger">Eliminar</button></td>
-                </td>
+                    </td>
 
-                <td>
-                    <a href='editar/<?php echo $_smarty_tpl->tpl_vars['brand']->value->{$_smarty_tpl->tpl_vars['id_brand']->value};?>
+                    <td>
+                        <a href='editar/<?php echo $_smarty_tpl->tpl_vars['brand']->value->{$_smarty_tpl->tpl_vars['id_brand']->value};?>
 ' type='button' class="btn btn-success">Editar</button>
-                </td>
-
+                    </td>
+                <?php }?>
             </tr>
         <?php ob_start();
 }
